Add locale check for zsh
If locale settings aren't set to UTF-8 in zsh, set them.
This is usually covered by Terminal.app but Prompt doesn't so adding this
check in manually. Can remove in the future if Prompt does it correctly
(or it will just be ignored).

Need to confirm this works fine on Linux.