Enable hxtheme helper
hxtheme sets and reloads theme in helix. `config.toml` should be set to use `auto`. `hxtheme` will write an `auto.toml` that inherits the chosen them. This is useful for setting the theme on the command line and can be used with a trigger for when a system moves from light to dark mode to change the theme.