Terminal theme for Helix
`Terminal` is a theme that is optimized for the Apple dynamic color system used to support light/dark modes. In particular, the background, text, and accent colors adapt to whether the system is in light or dark mode.

Syntax coloring is limited to text effects.

`Terminal Light` has a couple modifications to fix the Terminal theme for strictly light palette (like Terminal.app's `Basic). Dark themes should work with Terminal but it is optimized for macOS Developer palette.

Removed other unused themes.